Get the server code under control
