+++ src/server/index.ts
@@ -1,12 +1,12 @@
 import { Router } from 'itty-router';
 import './sentry';
-import {WeightedMovie} from './handlers/Scored';
+import {ScoredMovie} from './handlers/ScoredMovie';
 import {MovieById} from './handlers/MovieById';
 
 const router = Router();
 
 router
-  .get('/api/movie', WeightedMovie)
+  .get('/api/movie', ScoredMovie)
   .get('/api/movie/:id', MovieById)
   .get('*', () => new Response('Not Found :(', { status: 404 }));
